pyproject.toml
[tool.poetry]
name = "image-compressor"
version = "0.1.0"
description = "compress images on folder"
authors = ["Guilherme Rodrigues <guilherme.b.rodrigues@skf.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
Pillow = "^9.1.1"

[tool.poetry.dev-dependencies]
blue = "^0.9.0"
isort = "^5.10.1"
pyinstaller = "^5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
